requirements.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
pywhatkit==5.4
wikipedia-api==0.5.4
PyJokes==0.6.0
